error products api

+++ app/Http/Controllers/Api/ProductController.php
@@ -22,6 +22,7 @@ class ProductController extends BaseController
 
         return $this->sendResponse(ProductResource::collection($products), 'Products retrieved successfully.');
     }
+
     /**
      * Store a newly created resource in storage.
      *
@@ -30,18 +31,19 @@ class ProductController extends BaseController
      */
     public function store(Request $request): JsonResponse
     {
-        $input = $request->all();
-
-        $validator = Validator::make($input, [
+        $validator = Validator::make($request->all(), [
             'name' => 'required',
-            'detail' => 'required'
+            'detail' => 'required',
         ]);
 
-        if($validator->fails()){
+        if ($validator->fails()) {
             return $this->sendError('Validation Error.', $validator->errors());
         }
 
-        $product = Product::create($input);
+        $product = new Product();
+        $product->name = $request->input('name');
+        $product->detail = $request->input('detail');
+        $product->save();
 
         return $this->sendResponse(new ProductResource($product), 'Product created successfully.');
     }
@@ -70,21 +72,25 @@ class ProductController extends BaseController
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Product $product): JsonResponse
+    public function update(Request $request, $id): JsonResponse
     {
-        $input = $request->all();
-
-        $validator = Validator::make($input, [
+        $validator = Validator::make($request->all(), [
             'name' => 'required',
-            'detail' => 'required'
+            'detail' => 'required',
         ]);
 
-        if($validator->fails()){
+        if ($validator->fails()) {
             return $this->sendError('Validation Error.', $validator->errors());
         }
 
-        $product->name = $input['name'];
-        $product->detail = $input['detail'];
+        $product = Product::find($id);
+
+        if (is_null($product)) {
+            return $this->sendError('Product not found.');
+        }
+
+        $product->name = $request->input('name');
+        $product->detail = $request->input('detail');
         $product->save();
 
         return $this->sendResponse(new ProductResource($product), 'Product updated successfully.');
@@ -96,8 +102,14 @@ class ProductController extends BaseController
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Product $product): JsonResponse
+    public function destroy($id): JsonResponse
     {
+        $product = Product::find($id);
+
+        if (is_null($product)) {
+            return $this->sendError('Product not found.');
+        }
+
         $product->delete();
 
         return $this->sendResponse([], 'Product deleted successfully.');
